backend/pyproject.toml
[tool.poetry]
name = "socratic"
version = "0.1.0"
description = "TODO"
authors = []

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.115.11"
gunicorn = "^23.0.0"
uvicorn = "^0.34.0"

mako = "~1.2.2"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"

alembic = "^1.14.1"
SQLAlchemy = "^2.0.38"
asyncpg = "^0.30.0"  # DB backend for application
psycopg2-binary = "~2.9.9"  # DB backend for alembic (migration tool)

redis = "^5.2.1"
saq = { version = "~0.12.4", extras = ["hiredis"] }

importlib-resources = { version = "~5.4.0", python = "<3.9" }
zipp = { version = "~3.7.0", python = "<3.9" }
importlib-metadata = { version = "~4.11.0", python = "<3.9" }
python-jose = "^3.4.0"
cryptography = "^44.0.2"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
email-validator = "^2.2.0"
httpx = "^0.28.1"
greenlet = "^3.1.1"
python-multipart = "^0.0.20"



[tool.poetry.dev-dependencies]
add-trailing-comma = "2.2.1"
autoflake = "1.4.0"
black = "22.3.0"
isort = "5.10.1"
mypy = "1.0.1"
wemake-python-styleguide = "0.16.0"

flake8-bandit = "2.1.2"  # https://github.com/PyCQA/bandit/issues/837
bandit = "1.7.2"  # https://github.com/PyCQA/bandit/issues/837

pytest = "^8.2.0"
pytest-asyncio = "^0.25.3"
pytest-cov = "~3.0.0"

asgi-lifespan = "~1.0.1"
requests = "~2.31.0"
respx = "~0.20.0"

markdown = "3.3.6"  # https://github.com/python-poetry/poetry/issues/4777

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"
